requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
topsis-mrinank-102303235==0.0.1
requests==2.31.0
